refactor: extract useVerifyLatestUser and wrap just app with authBoundary
