+++ adapter/src/components/AppWrapper.js
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 import { useCurrentUserLocale } from '../utils/useLocale.js'
+import { useVerifyLatestUser } from '../utils/useVerifyLatestUser.js'
 import { Alerts } from './Alerts.js'
 import { AuthBoundary } from './AuthBoundary.js'
 import { ConnectedHeaderBar } from './ConnectedHeaderBar.js'
@@ -10,24 +11,23 @@ import { styles } from './styles/AppWrapper.style.js'
 
 export const AppWrapper = ({ children }) => {
     const { loading: localeLoading } = useCurrentUserLocale()
+    const { loading: latestUserLoading, user } = useVerifyLatestUser()
 
-    if (localeLoading) {
+    if (localeLoading || latestUserLoading) {
         return <LoadingMask />
     }
 
     return (
-        <AuthBoundary>
-            <div className="app-shell-adapter">
-                <style jsx>{styles}</style>
-                <ConnectedHeaderBar />
-                <div className="app-shell-app">
-                    <ErrorBoundary onRetry={() => window.location.reload()}>
-                        {children}
-                    </ErrorBoundary>
-                </div>
-                <Alerts />
+        <div className="app-shell-adapter">
+            <style jsx>{styles}</style>
+            <ConnectedHeaderBar />
+            <div className="app-shell-app">
+                <ErrorBoundary onRetry={() => window.location.reload()}>
+                    <AuthBoundary user={user}>{children}</AuthBoundary>
+                </ErrorBoundary>
             </div>
-        </AuthBoundary>
+            <Alerts />
+        </div>
     )
 }
